Add container serialization.
