+++ modules/serdes/src/protobuf/containers.cpp
@@ -1,0 +1,5 @@
+//=================================================================================================
+// Copyright (C) 2023-2024 HEPHAESTUS Contributors
+//=================================================================================================
+
+#include "hephaestus/serdes/protobuf/containers.h"
