+++ modules/types_proto/src/dummy_type.cpp
@@ -7,33 +7,11 @@
 #include <cstddef>
 #include <vector>
 
+#include "hephaestus/serdes/protobuf/containers.h"
 #include "hephaestus/types/dummy_type.h"
 #include "hephaestus/types/proto/dummy_type.pb.h"
 
 namespace heph::types {
-
-//=================================================================================================
-// Vector
-//=================================================================================================
-template <typename T, typename ProtoT>
-auto toProto(google::protobuf::RepeatedField<ProtoT>& proto_repeated_field, const std::vector<T>& vec)
-    -> void {
-  proto_repeated_field.Clear();  // Ensure that the repeated field is empty before adding elements.
-  proto_repeated_field.Reserve(static_cast<int>(vec.size()));
-  for (const auto& value : vec) {
-    proto_repeated_field.Add(value);
-  }
-}
-
-template <typename T, typename ProtoT>
-auto fromProto(const google::protobuf::RepeatedField<ProtoT>& proto_repeated_field, std::vector<T>& vec)
-    -> void {
-  vec.clear();  // Ensure that the vector is empty before adding elements.
-  vec.reserve(static_cast<size_t>(proto_repeated_field.size()));
-  for (const auto& proto_value : proto_repeated_field) {
-    vec.push_back(proto_value);
-  }
-}
 
 //=================================================================================================
 // DummyTypeDummyEnum
@@ -133,7 +111,7 @@ auto toProto(proto::DummyType& proto_dummy_type, const DummyType& dummy_type) ->
 
   proto_dummy_type.set_dummy_string(dummy_type.dummy_string);
 
-  toProto(*proto_dummy_type.mutable_dummy_vector(), dummy_type.dummy_vector);
+  serdes::protobuf::toProto(*proto_dummy_type.mutable_dummy_vector(), dummy_type.dummy_vector);
 }
 
 auto fromProto(const proto::DummyType& proto_dummy_type, DummyType& dummy_type) -> void {
@@ -143,7 +121,7 @@ auto fromProto(const proto::DummyType& proto_dummy_type, DummyType& dummy_type) 
 
   dummy_type.dummy_string = proto_dummy_type.dummy_string();
 
-  fromProto(proto_dummy_type.dummy_vector(), dummy_type.dummy_vector);
+  serdes::protobuf::fromProto(proto_dummy_type.dummy_vector(), dummy_type.dummy_vector);
 }
 
 }  // namespace heph::types
